Last commit in a while. Made lots of changes to the setup screen, just about to add edit year and done button functionality.

+++ NCEACredits/Year.m
@@ -43,7 +43,7 @@
 //****
 //*
 
-+ (NSUInteger)getCurrentYearDate {
++ (unsigned long)getCurrentYearDate {
     NSDateComponents *comp = [[NSCalendar currentCalendar]
                               components: NSCalendarUnitYear
                               fromDate:[NSDate date]];
